+++ docs/CHANGES.md
@@ -1,6 +1,7 @@
 2017-02-10  Release 1.77.1
 
   * Color the report date yellow or red, depending on the age of the report.
+  * Bug fix: Don't crash when the coverage percentage in a NCover report is missing.
   
 
 2017-02-09  Release 1.77.0